Add files via upload
Designed a Finite State Automaton(FSA) meant to simulate a washing machine. Specifically focused on implementing state and memento design patterns. The design included an FSA class that kept track of the current setting of the washing machine and the set of all its possible settings. This class simulates setting changes of the washing machine by changing the state based on an input(button press). Created a state class that handles the creation of the states used in the FSA and includes methods to handle entry, steady-state, and exit behaviors. Created a memento class that saves the settings of the FSA that correspond to the settings of the machine. This class lets you revert back to any previous settings even if one couldn't directly go to that setting.